fix: Somebody has been 🧑‍🦯 and switched `importantstuff-db-client` and `studentcheckbox-db-client` (and it's not me)
